+++ Sources/ShopifyCheckoutSheetKit/Configuration.swift
@@ -39,11 +39,16 @@ public struct Configuration {
 
 	public var preloading = Configuration.Preloading()
 
+	public var tintColor: UIColor = UIColor(red: 0.09, green: 0.45, blue: 0.69, alpha: 1.00)
+
+	@available(*, renamed: "tintColor", message: "spinnerColor has been superseded by tintColor")
 	public var spinnerColor: UIColor = UIColor(red: 0.09, green: 0.45, blue: 0.69, alpha: 1.00)
 
 	public var backgroundColor: UIColor = .systemBackground
 
 	public var logger: Logger = NoOpLogger()
+
+	public var title: String = NSLocalizedString("shopify_checkout_sheet_title", value: "Checkout", comment: "The title of the checkout sheet.")
 }
 
 extension Configuration {